Resupply Works
Resupply works but is buggy, will have to jump on the side of the 2 lower boxes and look at the top box, press E you can buy the ammo but the player will shoot. you can reload or shoot again and the ammo updates. I gotta run to work im late....